% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almond.R
\name{almondPage}
\alias{almondPage}
\title{Create a page with Almond.CSS}
\source{
<https://github.com/alvaromontoro/almond.css>
}
\usage{
almondPage(..., title = NULL, lang = NULL)
}
\arguments{
\item{...}{The contents of the document body}

\item{title}{The browser window title (defaults to the host URL of the page)}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (\code{NULL}) results in an empty string}
}
\description{
Create a Shiny UI page that loads the CSS for Almond.CSS,
and has no content in the page body (other than what you provide).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(classless)

ui <- almondPage(
  h1("Welcome to Almond.CSS!"),
  p("A classless CSS framework")
)

server <- function(input, output) {}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
\seealso{
AlmondCSS
}
