% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almond.R
\name{almondTheme}
\alias{almondTheme}
\title{Add Almond.CSS Theme to UI}
\usage{
almondTheme(lite = FALSE)
}
\arguments{
\item{lite}{Logical, should the lite version of almond.css be used? Default is \code{FALSE}}
}
\value{
An \code{\link[htmltools]{htmlDependency}} to the relevant Almond.CSS style file
}
\description{
Creates the dependency required for Almond.CSS to be included in the HTML of a page.
}
\examples{
library(htmltools)
library(classless)

ui <- tags$html()
ui <- htmltools::attachDependencies(
  ui,
  almondTheme()
)

}
