% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{actionCLButton}
\alias{actionCLButton}
\alias{actionCLLink}
\title{Action Button/Link}
\usage{
actionCLButton(input_id, label, ...)

actionCLLink(input_id, label, ...)
}
\arguments{
\item{input_id}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label, but you could also use any other HTML, like an image.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Creates an action button or link whose value is initially zero, and increments by one each time it is pressed.
}
\examples{
actionCLButton("button", "Click Me")

}
