% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_classless.R
\name{ClasslessCSS}
\alias{ClasslessCSS}
\title{Classless.CSS}
\description{
Classless.CSS

Classless.CSS
}
\seealso{
\code{\link{classlessPage}}, \code{\link{listClasslessThemes}}
}
\section{Super class}{
\code{\link[classless:Classless]{classless::Classless}} -> \code{ClasslessCSS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClasslessCSS$new()}}
\item \href{#method-createPage}{\code{ClasslessCSS$createPage()}}
\item \href{#method-clone}{\code{ClasslessCSS$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="addServer">}\href{../../classless/html/Classless.html#method-addServer}{\code{classless::Classless$addServer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="addUI">}\href{../../classless/html/Classless.html#method-addUI}{\code{classless::Classless$addUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="dateInput">}\href{../../classless/html/Classless.html#method-dateInput}{\code{classless::Classless$dateInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getFramework">}\href{../../classless/html/Classless.html#method-getFramework}{\code{classless::Classless$getFramework()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getServer">}\href{../../classless/html/Classless.html#method-getServer}{\code{classless::Classless$getServer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getUI">}\href{../../classless/html/Classless.html#method-getUI}{\code{classless::Classless$getUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="numericInput">}\href{../../classless/html/Classless.html#method-numericInput}{\code{classless::Classless$numericInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="passwordInput">}\href{../../classless/html/Classless.html#method-passwordInput}{\code{classless::Classless$passwordInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="runApp">}\href{../../classless/html/Classless.html#method-runApp}{\code{classless::Classless$runApp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="runExample">}\href{../../classless/html/Classless.html#method-runExample}{\code{classless::Classless$runExample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="selectInput">}\href{../../classless/html/Classless.html#method-selectInput}{\code{classless::Classless$selectInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="textAreaInput">}\href{../../classless/html/Classless.html#method-textAreaInput}{\code{classless::Classless$textAreaInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="textInput">}\href{../../classless/html/Classless.html#method-textInput}{\code{classless::Classless$textInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Classless.CSS object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClasslessCSS$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ClasslessCSS} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPage"></a>}}
\if{latex}{\out{\hypertarget{method-createPage}{}}}
\subsection{Method \code{createPage()}}{
Create a page with Classless.CSS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClasslessCSS$createPage(..., title = NULL, theme = NULL, lang = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The contents of the document body}

\item{\code{title}}{The browser window title (defaults to the host URL of the page)}

\item{\code{theme}}{Theme to give to the UI}

\item{\code{lang}}{ISO 639-1 language code for the HTML page}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClasslessCSS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
