% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classless.R
\name{classlessPage}
\alias{classlessPage}
\title{Create a page with Classless.css}
\source{
<https://github.com/emareg/classlesscss>
}
\usage{
classlessPage(..., title = NULL, theme = NULL, lang = NULL)
}
\arguments{
\item{...}{The contents of the document body}

\item{title}{The browser window title (defaults to the host URL of the page)}

\item{theme}{One of the following:
\itemize{
\item{\code{NULL} (the default), which implies a "stock" build of Classless.css}
\item{One of the themes available in \code{\link{listClasslessThemes}()}}
\item{A **named** character string pointing to an alternative Classless.css stylesheet
      (normally a css file within the www directory, e.g. \code{c(theme = "www/classless-theme.css")})}
}}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (\code{NULL}) results in an empty string}
}
\description{
Create a Shiny UI page that loads the CSS for Classless.css,
and has no content in the page body (other than what you provide).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(classless)

ui <- classlessPage(
  h1("Welcome to Classless!"),
  p("An (almost) classless CSS framework")
)

server <- function(input, output) {}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
