% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\name{getSakuraThemesheet}
\alias{getSakuraThemesheet}
\title{Get Sakura Themesheet}
\usage{
getSakuraThemesheet(file = "", append = TRUE)
}
\arguments{
\item{file}{Location to write the file to, or \code{NULL} to print to the console}

\item{append}{If \code{TRUE} output will be appended to file; otherwise, it will overwrite the contents of file}
}
\value{
A list of unasssigned style variables to be updated by the developer
}
\description{
This will write the required style variables to a specified location to customise the sakura theme.
}
\examples{
getSakuraThemesheet()

}
