% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classless.R
\name{listClasslessThemes}
\alias{listClasslessThemes}
\title{List Classless Themes}
\usage{
listClasslessThemes()
}
\value{
A vector of available CSS styles
}
\description{
Classless comes with a several themes that can be applied to the SCSS. This provides a list of all available options.
}
\examples{
listClasslessThemes()

}
