% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\name{listSakuraThemes}
\alias{listSakuraThemes}
\title{List Sakura Themes}
\usage{
listSakuraThemes()
}
\value{
A vector of available CSS styles
}
\description{
Sakura comes with a several themes that can be applied to the SCSS. This provides a list of all available options.
}
\examples{
listSakuraThemes()

}
