% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{runClasslessExample}
\alias{runClasslessExample}
\alias{runClasslessExample.default}
\alias{runClasslessExample.Classless}
\title{Run classless Example App}
\usage{
runClasslessExample(framework, ...)

\method{runClasslessExample}{default}(framework = FRAMEWORKS, ...)

\method{runClasslessExample}{Classless}(framework, ...)
}
\arguments{
\item{framework}{The name of the framework to}

\item{...}{Arguments to pass to the relevant page}
}
\description{
Test out the classless CSS frameworks. It is a kitchen sink of all of the HTML tags available.

All applications use the same R code, with the exception of the outer "page" function call.
}
