% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\name{sakuraPage}
\alias{sakuraPage}
\title{Create a page with Sakura}
\source{
<https://github.com/oxalorg/sakura>
}
\usage{
sakuraPage(..., title = NULL, theme = NULL, lang = NULL)
}
\arguments{
\item{...}{The contents of the document body}

\item{title}{The browser window title (defaults to the host URL of the page)}

\item{theme}{One of the following:
\itemize{
\item{\code{NULL} (the default), which implies a "stock" build of Sakura}
\item{One of the themes available in \code{\link{listSakuraThemes}()}}
\item{A character string pointing to an alternative Sakura stylesheet
      (normally a scss file within the www directory, e.g. www/sakura.scss)}
}}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (\code{NULL}) results in an empty string}
}
\description{
Create a Shiny UI page that loads the CSS for Sakura,
and has no content in the page body (other than what you provide).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(classless)

ui <- sakuraPage(
  h1("Welcome to Sakura!"),
  p("A classless CSS framework")
)

server <- function(input, output) {}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
