% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\name{sakuraTheme}
\alias{sakuraTheme}
\title{Add Sakura Theme to UI}
\usage{
sakuraTheme(
  theme = NULL,
  sass_options = sass::sass_options_get(output_style = "compressed"),
  cache = sass::sass_cache_get()
)
}
\arguments{
\item{theme}{One of the following:
\itemize{
\item{\code{NULL} (the default), which implies a "stock" build of Sakura}
\item{One of the themes available in \code{\link{listSakuraThemes}()}}
\item{A character string pointing to an alternative Sakura stylesheet
      (normally a scss file within the www directory, e.g. www/sakura.scss)}
}}

\item{sass_options}{A \code{\link[sass]{sass_options}()} object.}

\item{cache}{This can be a directory to use for the cache, a \code{\link[sass]{FileCache}}
object created by \code{\link[sass]{sass_file_cache}()}, or \code{FALSE} or \code{NULL} for no caching.}
}
\value{
An \code{\link[htmltools]{htmlDependency}} to the relevant Sakura style file
}
\description{
Creates the dependency required for Sakura to be included in the HTML of a page.
}
\examples{
library(htmltools)
library(classless)

ui <- tags$html()
ui <- htmltools::attachDependencies(
  ui,
  sakuraTheme()
)

}
