% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_almond.R
\name{AlmondCSS}
\alias{AlmondCSS}
\title{Almond.CSS}
\source{
<https://github.com/alvaromontoro/almond.css>
}
\description{
A ...
}
\examples{
library(shiny)
library(classless)

almond <- AlmondCSS$new()

ui <- almond$createPage(
  tags$header(
    "Almond Application"
  ),
  tags$body(
    tags$section(
      tags$h2("Contents"),
      tags$p("Paragraph")
    ),
    tags$section(
      tags$h2("Inputs"),
      tags$form(
        almond$textInput("text", "Text Input"),
        almond$numericInput("number", "Numeric Input")
      )
    )
  )
)

if (interactive()) {
  server <- function(input, output, session) {}
  shinyApp(ui, server)
}

}
\seealso{
\code{\link{almondPage}}

selectCLInput
Convert Classless Input to Almond.css Input
}
\section{Super class}{
\code{\link[classless:Classless]{classless::Classless}} -> \code{AlmondCSS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AlmondCSS$new()}}
\item \href{#method-createPage}{\code{AlmondCSS$createPage()}}
\item \href{#method-textInput}{\code{AlmondCSS$textInput()}}
\item \href{#method-passwordInput}{\code{AlmondCSS$passwordInput()}}
\item \href{#method-textAreaInput}{\code{AlmondCSS$textAreaInput()}}
\item \href{#method-numericInput}{\code{AlmondCSS$numericInput()}}
\item \href{#method-dateInput}{\code{AlmondCSS$dateInput()}}
\item \href{#method-selectInput}{\code{AlmondCSS$selectInput()}}
\item \href{#method-almondInput}{\code{AlmondCSS$almondInput()}}
\item \href{#method-clone}{\code{AlmondCSS$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="actionButton">}\href{../../classless/html/Classless.html#method-actionButton}{\code{classless::Classless$actionButton()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="addServer">}\href{../../classless/html/Classless.html#method-addServer}{\code{classless::Classless$addServer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="addUI">}\href{../../classless/html/Classless.html#method-addUI}{\code{classless::Classless$addUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="checkboxInput">}\href{../../classless/html/Classless.html#method-checkboxInput}{\code{classless::Classless$checkboxInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getFramework">}\href{../../classless/html/Classless.html#method-getFramework}{\code{classless::Classless$getFramework()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getServer">}\href{../../classless/html/Classless.html#method-getServer}{\code{classless::Classless$getServer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getUI">}\href{../../classless/html/Classless.html#method-getUI}{\code{classless::Classless$getUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="runApp">}\href{../../classless/html/Classless.html#method-runApp}{\code{classless::Classless$runApp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="runExample">}\href{../../classless/html/Classless.html#method-runExample}{\code{classless::Classless$runExample()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Almond.CSS object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{AlmondCSS} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPage"></a>}}
\if{latex}{\out{\hypertarget{method-createPage}{}}}
\subsection{Method \code{createPage()}}{
Create a page with Almond.CSS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$createPage(..., title = NULL, lang = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The contents of the document body}

\item{\code{title}}{The browser window title (defaults to the host URL of the page)}

\item{\code{lang}}{ISO 639-1 language code for the HTML page}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput"></a>}}
\if{latex}{\out{\hypertarget{method-textInput}{}}}
\subsection{Method \code{textInput()}}{
Create an input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$textInput(
  input_id,
  label,
  value = "",
  placeholder = "",
  inline = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A text input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-passwordInput"></a>}}
\if{latex}{\out{\hypertarget{method-passwordInput}{}}}
\subsection{Method \code{passwordInput()}}{
Create an password control for entry of passwords
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$passwordInput(
  input_id,
  label,
  value = "",
  placeholder = "",
  inline = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A password input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textAreaInput"></a>}}
\if{latex}{\out{\hypertarget{method-textAreaInput}{}}}
\subsection{Method \code{textAreaInput()}}{
Create a textarea input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$textAreaInput(
  input_id,
  label,
  value = "",
  placeholder = "",
  inline = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A textarea input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput}{}}}
\subsection{Method \code{numericInput()}}{
Create a numeric input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$numericInput(
  input_id,
  label,
  value = "",
  min = NULL,
  max = NULL,
  placeholder = "",
  inline = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{min}}{Minimum valid value}

\item{\code{max}}{Maximum valid value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dateInput"></a>}}
\if{latex}{\out{\hypertarget{method-dateInput}{}}}
\subsection{Method \code{dateInput()}}{
Create a date input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$dateInput(
  input_id,
  label,
  value = "",
  min = NULL,
  max = NULL,
  placeholder = "",
  inline = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{min}}{Minimum valid value}

\item{\code{max}}{Maximum valid value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A date input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectInput"></a>}}
\if{latex}{\out{\hypertarget{method-selectInput}{}}}
\subsection{Method \code{selectInput()}}{
Create a select list that can be used to choose a single or multiple items from a list of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$selectInput(
  input_id,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  inline = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{choices}}{List of values to select from}

\item{\code{selected}}{The initially selected value (or multiple values if \code{multiple = TRUE}).}

\item{\code{multiple}}{Is selection of multiple items allowed?}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A select list control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-almondInput"></a>}}
\if{latex}{\out{\hypertarget{method-almondInput}{}}}
\subsection{Method \code{almondInput()}}{
Almond.css adds a couple of tags to inputs that some classless inputs don't have
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$almondInput(input, inline = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Some UI input}

\item{\code{inline}}{Logical, whether the component is in-line with sibling elements}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlmondCSS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
