% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sakura.R
\name{SakuraCSS}
\alias{SakuraCSS}
\title{Sakura.CSS}
\description{
A ...
}
\seealso{
\code{\link{sakuraPage}}, \code{\link{listSakuraThemes}}
}
\section{Super class}{
\code{\link[classless:Classless]{classless::Classless}} -> \code{SakuraCSS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SakuraCSS$new()}}
\item \href{#method-createPage}{\code{SakuraCSS$createPage()}}
\item \href{#method-textInput}{\code{SakuraCSS$textInput()}}
\item \href{#method-passwordInput}{\code{SakuraCSS$passwordInput()}}
\item \href{#method-textAreaInput}{\code{SakuraCSS$textAreaInput()}}
\item \href{#method-numericInput}{\code{SakuraCSS$numericInput()}}
\item \href{#method-dateInput}{\code{SakuraCSS$dateInput()}}
\item \href{#method-sakuraInput}{\code{SakuraCSS$sakuraInput()}}
\item \href{#method-clone}{\code{SakuraCSS$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="getFramework">}\href{../../classless/html/Classless.html#method-getFramework}{\code{classless::Classless$getFramework()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="classless" data-topic="Classless" data-id="runExample">}\href{../../classless/html/Classless.html#method-runExample}{\code{classless::Classless$runExample()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new sakura.css object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{SakuraCSS} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPage"></a>}}
\if{latex}{\out{\hypertarget{method-createPage}{}}}
\subsection{Method \code{createPage()}}{
Create a page with sakura.css
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$createPage(..., title = NULL, theme = NULL, lang = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The contents of the document body}

\item{\code{title}}{The browser window title (defaults to the host URL of the page)}

\item{\code{theme}}{Theme to give to the UI}

\item{\code{lang}}{ISO 639-1 language code for the HTML page}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput"></a>}}
\if{latex}{\out{\hypertarget{method-textInput}{}}}
\subsection{Method \code{textInput()}}{
Create an input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$textInput(input_id, label, value = "", placeholder = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A text input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-passwordInput"></a>}}
\if{latex}{\out{\hypertarget{method-passwordInput}{}}}
\subsection{Method \code{passwordInput()}}{
Create an password control for entry of passwords
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$passwordInput(input_id, label, value = "", placeholder = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A password input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textAreaInput"></a>}}
\if{latex}{\out{\hypertarget{method-textAreaInput}{}}}
\subsection{Method \code{textAreaInput()}}{
Create a textarea input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$textAreaInput(input_id, label, value = "", placeholder = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A textarea input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput}{}}}
\subsection{Method \code{numericInput()}}{
Create a numeric input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$numericInput(
  input_id,
  label,
  value = "",
  min = NULL,
  max = NULL,
  placeholder = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{min}}{Minimum valid value}

\item{\code{max}}{Maximum valid value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric input control that can be added to a UI definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dateInput"></a>}}
\if{latex}{\out{\hypertarget{method-dateInput}{}}}
\subsection{Method \code{dateInput()}}{
Create a date input control for entry of unstructured text values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$dateInput(
  input_id,
  label,
  value = "",
  min = NULL,
  max = NULL,
  placeholder = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_id}}{The \code{input} slot that will be used to access the value.}

\item{\code{label}}{Display label for the control}

\item{\code{value}}{Initial value}

\item{\code{min}}{Minimum valid value}

\item{\code{max}}{Maximum valid value}

\item{\code{placeholder}}{A character string giving the user a hint as to what can be entered into the control}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A date input control that can be added to a UI definition
Convert Classless Input to sakura.css Input
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sakuraInput"></a>}}
\if{latex}{\out{\hypertarget{method-sakuraInput}{}}}
\subsection{Method \code{sakuraInput()}}{
sakura.css wraps a \code{<p>} tag around classless form inputs to achieve the correct style
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$sakuraInput(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Some UI input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SakuraCSS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
